analyzer: Some algorithms fixes.

+++ analyser/src/org/ebooksearchtool/analyzer/algorithms/subalgorithms/SizeExtractor.java
@@ -40,12 +40,12 @@ public class SizeExtractor {
                 
                 if(lex.indexOf("KB") != -1){
                     lex = lex.substring(0, lex.length() - 2);
-                    long size = Integer.parseInt(lex);
+                    long size = (long) Double.parseDouble(lex);
                     size*=1024;
                     lex = String.valueOf(size);
                 }else{
                     lex = lex.substring(0, lex.length() - 2);
-                    long size = Integer.parseInt(lex);
+                    long size = (long) Double.parseDouble(lex);
                     size*=1024;
                     size*=1024;
                     lex = String.valueOf(size);
